@@ -3,29 +3,29 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@elastic/eui": "^26.1.0",
-    "@fullhuman/postcss-purgecss": "^2.2.0",
-    "@splidejs/react-splide": "^0.7.3",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "autoprefixer": "^9.8.0",
-    "firebase": "^7.14.6",
-    "html-to-react": "^1.4.8",
+    "@fullhuman/postcss-purgecss": "^5.0.0",
+    "@splidejs/react-splide": "^0.7.12",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
+    "autoprefixer": "^10.4.14",
+    "firebase": "^9.17.2",
+    "html-to-react": "^1.5.0",
     "japaneasy": "^0.2.2",
-    "node-fetch": "^3.2.3",
+    "node-fetch": "^3.3.1",
+    "npm-check-updates": "^16.7.12",
     "npm-run-all": "^4.1.5",
-    "postcss-cli": "^7.1.1",
-    "react": "^16.13.1",
-    "react-card-flip": "^1.0.10",
-    "react-dom": "^16.13.1",
-    "react-icons": "^4.3.1",
-    "react-parallax-tilt": "^1.3.42",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.1",
-    "react-transition-group": "^4.4.2",
-    "react-window": "^1.8.6",
-    "tailwindcss": "^1.4.6"
+    "postcss-cli": "^10.1.0",
+    "react": "^18.2.0",
+    "react-card-flip": "^1.2.0",
+    "react-dom": "^18.0.0",
+    "react-icons": "^4.8.0",
+    "react-parallax-tilt": "^1.7.115",
+    "react-router-dom": "^6.9.0",
+    "react-scripts": "5.0.1",
+    "react-transition-group": "^4.4.5",
+    "react-window": "^1.8.8",
+    "tailwindcss": "^3.2.7"
   },
   "scripts": {
     "build:css": "postcss src/styles/App.css -o src/index.css",
@@ -51,6 +51,6 @@
     ]
   },
   "devDependencies": {
-    "concurrently": "^5.2.0"
+    "concurrently": "^7.6.0"
   }
 }
