@@ -14,17 +14,15 @@ const Layout = ({children}) => {
     const windowsSize = useWindowSize()
 
     return (
-        <div className='h-full'>
-            <header id="header-nav" className='w-full h-24 flex justify-center'>
+        <>
             <Nav link={Link} />
-            </header>
             {children}
             <Footer/>
             {windowsSize.width < 560 && ( 
                     <HatsuNatsuMode/> 
             )}
             {isOpen ? <LearningListModal setIsOpen={setIsOpen} /> : <button className='z-50 list-btn text-white' onClick={() => setIsOpen(true)}>Liste</button>}
-      </div>  
+        </>
     )
 }
 
